RPC example -- one RPC call with response -- complete.
Note that a tool should be created which automatically creates the required client and server stubs.
This tool could also take care for creating the multiplexing of different RPC types.

+++ core/comm/test/rpc-sample-interface.hpp
@@ -8,7 +8,7 @@ using namespace std;
 
 //@serializable
 struct DataForRPC1{
-	int a;
+	int64_t a;
 };
 
 //@serializable
@@ -18,7 +18,7 @@ struct DataForRPC2{
 
 //@serializable
 struct DataFromRPC2{
-	int c;
+	int32_t c;
 	string d;
 };
